# Optimize concat for the WebGL backend