--- a/src/environment_util.ts
+++ b/src/environment_util.ts
@@ -42,6 +42,8 @@ export interface Features {
   'WEBGL_PAGING_ENABLED'?: boolean;
   // The maximum texture dimension.
   'WEBGL_MAX_TEXTURE_SIZE'?: number;
+  // The maximum number of textures in a single shader program.
+  'WEBGL_MAX_TEXTURES_IN_SHADER'?: number;
   // The disjoint_query_timer extension version.
   // 0: disabled, 1: EXT_disjoint_timer_query, 2:
   // EXT_disjoint_timer_query_webgl2.
@@ -101,6 +103,7 @@ export const URL_PROPERTIES: URLProperty[] = [
   {name: 'WEBGL_PACK_DEPTHWISECONV', type: Type.BOOLEAN},
   {name: 'WEBGL_CONV_IM2COL', type: Type.BOOLEAN},
   {name: 'WEBGL_MAX_TEXTURE_SIZE', type: Type.NUMBER},
+  {name: 'WEBGL_MAX_TEXTURES_IN_SHADER', type: Type.NUMBER},
   {name: 'WEBGL_PAGING_ENABLED', type: Type.BOOLEAN},
   {name: 'WEBGL_DISJOINT_QUERY_TIMER_EXTENSION_VERSION', type: Type.NUMBER},
   {name: 'WEBGL_DISJOINT_QUERY_TIMER_EXTENSION_RELIABLE', type: Type.BOOLEAN},
@@ -132,16 +135,26 @@ export function isWebGLVersionEnabled(webGLVersion: 1|2) {
   return false;
 }
 
-let MAX_TEXTURE_SIZE: number;
-// Caching MAX_TEXTURE_SIZE here because the environment gets reset between
+// We cache webgl params because the environment gets reset between
 // unit tests and we don't want to constantly query the WebGLContext for
 // MAX_TEXTURE_SIZE.
+let MAX_TEXTURE_SIZE: number;
+let MAX_TEXTURES_IN_SHADER: number;
+
 export function getWebGLMaxTextureSize(webGLVersion: number): number {
   if (MAX_TEXTURE_SIZE == null) {
     const gl = getWebGLContext(webGLVersion);
     MAX_TEXTURE_SIZE = gl.getParameter(gl.MAX_TEXTURE_SIZE);
   }
   return MAX_TEXTURE_SIZE;
+}
+
+export function getMaxTexturesInShader(webGLVersion: number): number {
+  if (MAX_TEXTURES_IN_SHADER == null) {
+    const gl = getWebGLContext(webGLVersion);
+    MAX_TEXTURES_IN_SHADER = gl.getParameter(gl.MAX_TEXTURE_IMAGE_UNITS);
+  }
+  return MAX_TEXTURES_IN_SHADER;
 }
 
 export function getWebGLDisjointQueryTimerVersion(webGLVersion: number):