@dsmilkov 2018-12-12
Good q. A quick search for (`Tensor[]` or `T[]`) in backend.ts gives `addN()` as another candidate.